chore(20006): move helper functions for layer fatures atom into separate folder
